+++ src/Pgsql/InsertBuilder.php
@@ -20,6 +20,7 @@ use Aura\SqlQuery\Common;
 class InsertBuilder extends Common\InsertBuilder
 {
     use BuildReturningTrait;
+    use Common\BuildOnConflictTrait;
 
     /**
      *