+++ src/Sqlite/Insert.php
@@ -9,6 +9,7 @@
 namespace Aura\SqlQuery\Sqlite;
 
 use Aura\SqlQuery\Common;
+use Aura\SqlQuery\Exception;
 
 /**
  *
@@ -17,9 +18,10 @@ use Aura\SqlQuery\Common;
  * @package Aura.SqlQuery
  *
  */
-class Insert extends Common\Insert
+class Insert extends Common\Insert implements Common\OnConflictUpdateInterface
 {
     use OrConflictTrait;
+    use Common\OnConflictUpdateTrait;
 
     /**
      *
@@ -30,8 +32,56 @@ class Insert extends Common\Insert
      */
     protected function build()
     {
+        $ignore = false;
+        $has_or_ignore = $this->hasFlag('OR IGNORE');
+        if (!empty($this->conflict_target)) {
+            if ($has_or_ignore) {
+                $this->setFlag('OR IGNORE', false);
+                $ignore = true;
+            }
+            $this->assertNoOrConflictFlags();
+        }
+
         $this->assertOneOrConflictFlag();
-        return parent::build();
+
+        $stm = parent::build();
+
+        if ($has_or_ignore && !empty($this->conflict_target)) {
+            $this->setFlag('OR IGNORE', true);
+        }
+
+        return $stm
+            . $this->builder->buildOnConflict(
+                $this->conflict_target,
+                $this->conflict_update_values,
+                $this->conflict_where,
+                $ignore
+            );
+    }
+
+    /**
+     *
+     * Asserts that no legacy SQLite OR conflict flags are set when using ON CONFLICT.
+     *
+     * @return void
+     * @throws Exception\LogicException
+     *
+     */
+    protected function assertNoOrConflictFlags()
+    {
+        $set = [];
+        foreach ($this->or_conflict_flags as $flag) {
+            if ($this->hasFlag($flag)) {
+                $set[] = $flag;
+            }
+        }
+
+        if (! empty($set)) {
+            throw new Exception\LogicException(
+                'Cannot combine ON CONFLICT clause with SQLite OR conflict flags: '
+                . implode(' and ', $set) . '.'
+            );
+        }
     }
 
     /**